move DisplayExpenseImage to ui/components/main package as separate component.
